Added neccessary Data structure for Security Implementation
I have added a beginning of the new database structure. This is essentially what I have presented earlier but includes the ideas and changes that were requested.

Changing the model requires first applying the model to a database, do your changes (including Foreign Key Constraints!) and then update the model.