Implemented Membership provider
I have implemented the basic membership provider we'll build upon. You may notice that there are a lot of stubs and NotImplementedExceptpions. However, this is totally fine because we won't use this stuff anyways. Most (if not all) of these stubs have become irrelevant over time and have been superseeded by EntityFramework (for Example: GetAllUsers) or do simply not apply to our Use case (GetNumberOfUsersOnline(), UpdateUser()). Tests are not done ;-)

Also there were some updates to the data storage.