[BVS] Notes, Advertisements, Multiaccounts, Roles
This commit adds a few Moderation tools:
- an user note module
- an advertisement module
- a multiaccount tracking module
- a role manager

While everything works (so far) it needs testing,
especially security-wise. Some GUIs are still far
from optimal (especially the role GUI, which
provides virtually no info and is kind of worth-
less without looking into the database).

Also added, because it was neccessary at some
point and helps debugging: Glimpse [0]. Glimpse
is a multi-purpose all-in-one debugging tool for
ASP.NET MVC and EntityFramework powered Webapps,
with route display and so on and so forth. Read
their webpage for more info. Remember it must be
enabled to function (visit /Glimpse.axd on every
subdomain).

Last, but not least, for initial setup you need
to register an account an manually give it at
least the "superadmin" and "moderator" roles.
Better give him all permissions.