[Util] Added dbtool to run migrations on GNU/Linux
I have added a small, simple and kind-of dirty tool to run migrations on
GNU/Linux. It is currently capable of applying the migrations to the database
as well as creating new migrations.

The interface is the same as in the Visual Studio CMDlets, however, to
successfully generate a migration, you need to specify the path where it should
be generated. This typically would be the directory of the other migrations.

Further help is available by just running ./dbtool.exe without arguments.